requirements.txt
Flask==3.0.0
google-cloud-storage==2.16.0
neo4j==5.15.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
setuptools>=65.0.0